Rename abstractions to ops.
