+++ fft/src/lib.rs
@@ -1,8 +1,8 @@
-pub mod abstractions;
 pub mod bit_reversing;
 pub mod errors;
 pub mod fft_iterative;
 pub(crate) mod helpers;
+pub mod ops;
 pub mod polynomial;
 pub mod roots_of_unity;
